fix(main): cast uint32_t draw-count to unsigned long for %lu format
gq_draw_oled_stack_count is a fixed-width uint32_t, but the exact-width
type is not guaranteed to be `unsigned int`. Format with %lu against an
explicit (unsigned long) cast, which is always a valid conversion target
per the C standard, instead of relying on %u matching the platform's
uint32_t typedef.

+++ gamequeer/src/main.c
@@ -55,7 +55,7 @@ static int dump_draw_count(const char *path) {
         fprintf(stderr, "dump_draw_count: cannot open %s for writing\n", path);
         return 0;
     }
-    fprintf(f, "%u\n", gq_draw_oled_stack_count);
+    fprintf(f, "%lu\n", (unsigned long) gq_draw_oled_stack_count);
     fclose(f);
     return 1;
 #else
